+++ R/remove_cash.R
@@ -1,0 +1,9 @@
+#' Remove cash
+#' @param data data.frame
+#' @param var variable name (tidy approach)
+remove_cash <- function(data, var) {
+  
+  data %>% 
+    dplyr::filter({{var}} != "Cash and/or Derivatives")
+  
+}