+++ tests/testthat/test-is_sparindex.R
@@ -1,0 +1,5 @@
+test_that("Test: is_sparindex", {
+  
+  expect_true(is_sparindex("Sparindex INDEX USA Growth KL"))
+  
+})